Try to fix problem on 64bit architectures (reported by Lars Schimmer). Hope that unsigned long int (ulong) is the type whose size equals the size of void * on every architecture. Please tell whether it helps.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@1748 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/game_debug_ui.cc
@@ -252,7 +252,7 @@ private:
 
 	UI::Multiline_Textarea*	m_ui_field;
 	UI::Button*					m_ui_immovable;
-	UI::Listselect<uint> * m_ui_bobs;
+	UI::Listselect<ulong> * m_ui_bobs;
 };
 
 
@@ -275,7 +275,7 @@ FieldDebugWindow::FieldDebugWindow(Interactive_Base* parent, Coords coords)
 	m_ui_immovable = new UI::Button(this, 0, 80, 200, 24, 0);
 	m_ui_immovable->clicked.set(this, &FieldDebugWindow::open_immovable);
 
-	m_ui_bobs = new UI::Listselect<uint>(this, 0, 104, 200, 96);
+	m_ui_bobs = new UI::Listselect<ulong>(this, 0, 104, 200, 96);
 	m_ui_bobs->selected.set(this, &FieldDebugWindow::open_bob);
 }
 
@@ -372,13 +372,10 @@ Open the bob debug window for the bob of the given index in the list
 ===============
 */
 void FieldDebugWindow::open_bob(const uint) {
-	if
-		(const unsigned long serial =
-		 static_cast<const unsigned long>(m_ui_bobs->get_selection()))
-		if
-			(Map_Object * const object =
-			 get_iabase()->get_egbase()->get_objects()->get_object(serial))
-			show_mapobject_debug(get_iabase(), object);
+	if (const ulong serial = m_ui_bobs->get_selection()) if
+		(Map_Object * const object =
+		 get_iabase()->get_egbase()->get_objects()->get_object(serial))
+		show_mapobject_debug(get_iabase(), object);
 }
